+++ packages/ui-new/src/Button/index.ts
@@ -1,2 +1,1 @@
 export * from "./Button";
-export * from "./ButtonPrimary";
