+++ packages/ui-new/src/Progress/index.ts
@@ -1,0 +1,1 @@
+export * from "./CircularProgress";
