--- packages/ui/src/Button/Button.styles.ts
@@ -1,14 +1,0 @@
-import { css } from "emotion";
-
-/**
- * Controls the button styles
- * @type {string}
- */
-const webinyButtonStyles = css`
-    &.webiny-ui-button--secondary:not(:disabled) {
-            border-color: var(--mdc-theme-primary);
-        }
-    }
-`;
-
-export { webinyButtonStyles };
