+++ packages/ui/src/Button/ButtonDefault.tsx
@@ -1,0 +1,6 @@
+import React from "react";
+import { Button, ButtonProps } from "@webiny/ui-new/Button";
+
+export const ButtonDefault = (props: ButtonProps) => {
+    return <Button {...props} variant={"primary"} />;
+};
