+++ packages/ui/src/Button/ButtonPrimary.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { Button, ButtonProps } from "./Button";
+import { Button, ButtonProps } from "@webiny/ui-new/Button";
 
 export const ButtonPrimary = (props: ButtonProps) => {
     return <Button {...props} variant={"primary"} />;
