+++ packages/ui/src/Button/ButtonSecondary.tsx
@@ -1,0 +1,6 @@
+import React from "react";
+import { Button, ButtonProps } from "@webiny/ui-new/Button";
+
+export const ButtonSecondary = (props: ButtonProps) => {
+    return <Button {...props} variant={"secondary"} />;
+};
