--- packages/ui/src/Button/CopyButton/README.md
@@ -1,11 +1,0 @@
-# Copy Button
-
-### Description
-
-Provides a button that allows you to copy its passed `value` to the clipboard.
-
-### Import
-
-```js
-import { CopyButton } from "@webiny/ui/Button";
-```
