--- packages/ui/src/Button/IconButton/IconButton.tsx
@@ -1,73 +1,0 @@
-import React from "react";
-import {
-    IconButton as RIconButton,
-    IconButtonProps as RmwcIconButtonProps
-} from "@rmwc/icon-button";
-import styled from "@emotion/styled";
-
-import { FormComponentProps } from "../../types";
-
-const StyledRIconButton = styled(RIconButton)`
-    i {
-        all: unset;
-        pointer-events: none;
-    }
-`;
-
-export interface IconButtonProps extends Omit<FormComponentProps, "onChange">, RmwcIconButtonProps {
-    id?: string;
-    /**
-     * Icon should be provided as an SvgComponent.
-     */
-    icon: React.ReactNode;
-
-    /**
-     * Button label
-     */
-    label?: string;
-
-    /**
-     * onClick handler
-     * @param event
-     */
-    onClick?: (event: React.MouseEvent) => void;
-
-    /**
-     * Custom CSS class
-     */
-    className?: string;
-    /**
-     * For testing purposes.
-     */
-
-    "data-testid"?: string;
-
-    /**
-     * Should icon be disabled?
-     */
-    disabled?: boolean;
-
-    children?: React.ReactNode;
-}
-
-/**
- * Shows the icon button.
- */
-const IconButton = (props: IconButtonProps) => {
-    const { id, icon, label, onClick, className, disabled, ripple } = props;
-
-    return (
-        <StyledRIconButton
-            id={id}
-            data-testid={props["data-testid"]}
-            onClick={onClick}
-            disabled={disabled}
-            className={className}
-            label={label}
-            icon={icon}
-            ripple={ripple}
-        />
-    );
-};
-
-export { IconButton };
