--- packages/ui/src/Button/IconButton/README.md
@@ -1,15 +1,0 @@
-# Icon Button
-
-### Design
-
-<a href="https://material.io/design/components/buttons.html" target="_blank">https://material.io/design/components/buttons.html</a>
-
-### Description
-
-`IconButton` component provides a simple icon button component with ripple effect.
-
-### Import
-
-```js
-import { IconButton } from "@webiny/ui/Button";
-```
