--- packages/ui/src/Button/README.md
@@ -1,28 +1,0 @@
-# Button
-
-### Design
-
-<a href="https://material.io/design/components/buttons.html" target="_blank">https://material.io/design/components/buttons.html</a>
-
-### Description
-
-`Button` component provides 4 commonly used button types:
-
-- `Default`
-- `Primary`
-- `Secondary`
-- `Floating`
-
-To create buttons with icon, use a `<Button.Icon/>` component.
-
-### Import
-
-```js
-import {
-  ButtonPrimary,
-  ButtonSecondary,
-  ButtonDefault,
-  ButtonFloating,
-  ButtonIcon
-} from "@webiny/ui/Button";
-```
