--- packages/ui/src/Progress/README.md
@@ -1,19 +1,0 @@
-# Progress
-
-### Design
-
-<a href="https://material.io/design/components/progress-indicators.html#circular-progress-indicators" target="_blank">
-    https://material.io/design/components/progress-indicators.html#circular-progress-indicators
-</a>
-
-### Description
-
-Progress components are used in cases where user has to wait for a background process to finish.
-
-Currently only the `CircularProgress` component is available, which is ideal for forms.
-
-### Import
-
-```js
-import { CircularProgress } from "@webiny/ui/Progress";
-```
